+++ contrib/fm_wavefront_tracker/__init__.py
@@ -5,11 +5,11 @@ executed when this submodule is imported, so adding things in this file can slow
 the importing process of `espresso`.
 
 For contributors, add any intialisation code for your problem into fm_wavefront_tracker.py, 
-under the method `__init__()` of the class `FmWavefrontTracker`.
+under the method `__init__()` of the class `FmmTomography`.
 
 Don't touch this file unless you know what you are doing :)
 """
 
 from .fm_wavefront_tracker import *
 
-__all__ = [ "FmWavefrontTracker" ]
+__all__ = [ "FmmTomography" ]