+++ contrib/fm_wavefront_tracker/fm_wavefront_tracker.py
@@ -7,7 +7,7 @@ from espresso.utils import absolute_path as path, silent_remove
 from . import waveTracker as wt
 
 
-class FmWavefrontTracker(EspressoProblem):
+class FmmTomography(EspressoProblem):
     """Forward simulation class
     """